+++ ohsome_quality_api/api/api.py
@@ -25,7 +25,8 @@ from ohsome_quality_api import (
     oqt,
 )
 from ohsome_quality_api.api.request_models import (
-    AttributeCompletenessRequest,
+    AttributeCompletenessFilterRequest,
+    AttributeCompletenessKeyRequest,
     IndicatorDataRequest,
     IndicatorRequest,
 )
@@ -268,12 +269,15 @@ async def post_indicator_ms(parameters: IndicatorDataRequest) -> CustomJSONRespo
 )
 async def post_attribute_completeness(
     request: Request,
-    parameters: AttributeCompletenessRequest,
+    parameters: AttributeCompletenessKeyRequest | AttributeCompletenessFilterRequest,
 ) -> Any:
     """Request the Attribute Completeness indicator for your area of interest."""
-    validate_attribute_topic_combination(
-        parameters.attribute_key.value, parameters.topic_key.value
-    )
+    if isinstance(parameters, AttributeCompletenessKeyRequest):
+        for attribute in parameters.attribute_keys:
+            validate_attribute_topic_combination(
+                attribute,
+                parameters.topic,
+            )
 
     return await _post_indicator(request, "attribute-completeness", parameters)
 
@@ -305,19 +309,12 @@ async def post_indicator(
 
 
 async def _post_indicator(
-    request: Request, key: str, parameters: IndicatorRequest
+    request: Request,
+    key: str,
+    parameters: IndicatorRequest,
 ) -> Any:
-    validate_indicator_topic_combination(key, parameters.topic_key.value)
-    attribute_key = getattr(parameters, "attribute_key", None)
-    if attribute_key:
-        attribute_key = attribute_key.value
-    indicators = await oqt.create_indicator(
-        key=key,
-        bpolys=parameters.bpolys,
-        topic=get_topic_preset(parameters.topic_key.value),
-        include_figure=parameters.include_figure,
-        attribute_key=attribute_key,
-    )
+    validate_indicator_topic_combination(key, parameters.topic)
+    indicators = await oqt.create_indicator(key=key, **dict(parameters))
 
     if request.headers["accept"] == MEDIA_TYPE_JSON:
         return {
@@ -338,10 +335,12 @@ async def _post_indicator(
         }
     else:
         detail = "Content-Type needs to be either {0} or {1}".format(
-            MEDIA_TYPE_JSON, MEDIA_TYPE_GEOJSON
+            MEDIA_TYPE_JSON,
+            MEDIA_TYPE_GEOJSON,
         )
         raise HTTPException(
-            status_code=status.HTTP_415_UNSUPPORTED_MEDIA_TYPE, detail=detail
+            status_code=status.HTTP_415_UNSUPPORTED_MEDIA_TYPE,
+            detail=detail,
         )