+++ ohsome_quality_api/api/request_models.py
@@ -1,12 +1,17 @@
-from typing import Dict
+from typing import Dict, List
 
 import geojson
 from geojson_pydantic import Feature, FeatureCollection, MultiPolygon, Polygon
-from pydantic import BaseModel, ConfigDict, Field, field_validator
+from pydantic import (
+    BaseModel,
+    ConfigDict,
+    Field,
+    field_validator,
+)
 
 from ohsome_quality_api.attributes.definitions import AttributeEnum
-from ohsome_quality_api.topics.definitions import TopicEnum
-from ohsome_quality_api.topics.models import TopicData
+from ohsome_quality_api.topics.definitions import TopicEnum, get_topic_preset
+from ohsome_quality_api.topics.models import TopicData, TopicDefinition
 from ohsome_quality_api.utils.helper import snake_to_lower_camel
 
 
@@ -49,7 +54,7 @@ class BaseBpolys(BaseConfig):
 
     @field_validator("bpolys")
     @classmethod
-    def transform(cls, value) -> geojson.FeatureCollection:
+    def transform_bpolys(cls, value) -> geojson.FeatureCollection:
         # NOTE: `geojson_pydantic` library is used only for validation and openAPI-spec
         # generation. To avoid refactoring all code the FeatureCollection object of
         # the `geojson` library is still used every else.
@@ -57,19 +62,44 @@ class BaseBpolys(BaseConfig):
 
 
 class IndicatorRequest(BaseBpolys):
-    topic_key: TopicEnum = Field(
+    topic: TopicEnum = Field(
         ...,
         title="Topic Key",
         alias="topic",
     )
     include_figure: bool = True
 
+    @field_validator("topic")
+    @classmethod
+    def transform_topic(cls, value) -> TopicDefinition:
+        return get_topic_preset(value.value)
 
-class AttributeCompletenessRequest(IndicatorRequest):
-    attribute_key: AttributeEnum = Field(
+
+class AttributeCompletenessKeyRequest(IndicatorRequest):
+    attribute_keys: List[AttributeEnum] = Field(
         ...,
-        title="Attribute Key",
-        alias="attribute",
+        title="Attribute Keys",
+        alias="attributes",
+    )
+
+    @field_validator("attribute_keys")
+    @classmethod
+    def transform_attributes(cls, value) -> list[str]:
+        return [attribute.value for attribute in value]
+
+
+class AttributeCompletenessFilterRequest(IndicatorRequest):
+    attribute_filter: str = Field(
+        ...,
+        title="Attribute Filter",
+        description="ohsome filter query representing custom attributes.",
+    )
+    attribute_title: str = Field(
+        ...,
+        title="Attribute Title",
+        description=(
+            "Title describing the attributes represented by the Attribute Filter."
+        ),
     )