+++ CHANGELOG.md
@@ -1,9 +1,53 @@
 # Changelog
 
-## Current Main 
+## Current Main
+
+### New Features
+
+- attribute completeness: added new topic and topic - specific attributes and deleted non-functional attributes ([#850])
 
 
-## Release  1.6.1 
+[#850]: https://github.com/GIScience/ohsome-quality-api/issue/850
+
+
+## Release  1.8.2 
+
+- PR template: Hurl maintenance is mentioned and wording with respect to mergeability is clarified ([#860])
+
+### Bug Fixes
+
+- attribute-completeness: custom filters are now correctly combined with topic filter ([#865])
+
+
+[#860]: https://github.com/GIScience/ohsome-quality-api/issues/860
+[#865]: https://github.com/GIScience/ohsome-quality-api/pull/865
+
+
+## Release  1.8.1 
+
+- attribute-completeness: indicator for `attribute-completeness` is now included in core profile ([#856])
+
+[#856]: https://github.com/GIScience/ohsome-quality-api/issues/856
+
+## Release  1.8.0
+
+### New Features
+
+- Support custom attribute definition via ohsome filter query for the Attribute Completeness indicator ([#848])
+
+[#848]: https://github.com/GIScience/ohsome-quality-api/pull/848
+
+## Release  1.7.0
+
+### Breaking Changes
+
+- attribute-completeness: `attribute` parameter is renamed to `attributes` and is now a list of attributes ([#819])
+
+
+[#819]: https://github.com/GIScience/ohsome-quality-api/issues/819
+
+
+## Release  1.6.1
 
 ### Bug Fixes